==> figure_3/1.calculate_avg_publications.ipynb <==
# Calculating avg. publication count

Classifying which editors are normal, questionable, and suspicious requires us knowing the average number of papers an editor publishes in his or her journal in the 5-year window (or as long as someone remains on the editorial board, whichever is shorter) after he or she becomes and editor, and 5-year window before someone becomes an editor (or as long one is research-active, whichever is shorter).

Again, computations demonstrated on the sample set of editors.

import pandas as pd

## Load data

%%time
editor_career = pd.read_csv('../data/EditorCareerDiscipline.csv',
            sep='\t', memory_map=True,
            usecols=['NewAuthorId', 'Yfp', 'Ylp', 'Parent'],
            dtype={'NewAuthorId':int, 'Yfp':int, 'Ylp':int, 'Parent':int})

papers = pd.read_csv('../data/EditorPapers.csv',sep='\t',
                    dtype={'NewAuthorId':int,'PaperId':int,'Year':int,'JournalId':int})
papers.shape, papers.NewAuthorId.nunique()

elsevier_journals = pd.read_csv('../data/ElsevierJournals.csv', sep='\t',
                               usecols=['JournalId','issn'], dtype={'JournalId':int, 'issn':str})
assert(elsevier_journals.issn.apply(lambda x: len(x)==8).all())
elsevier_journals.shape

editorials = pd.read_csv('../data/Editorials.csv', sep='\t', dtype={'PaperId':int}).assign(Editorial = 1)

# Average paper count and NQS

def printStats(df, msg=''):
    print(msg, df.shape, df[['NewAuthorId','issn']].drop_duplicates().shape)
    return df

def getEditors(filename, window=5):
    # window: the number of years that we filter

    editors = (
        pd.read_csv(filename, sep='\t',
                      dtype={'issn':str,'NewAuthorId':int,'start_year':int,'end_year':int})
        
        .assign(Year0 = lambda df: df.start_year.apply(lambda x: x-1))
        .query(f'start_year <= {2018-(window-1)}') # year0 prior to 2014
        .merge(editor_career, on='NewAuthorId')

        .assign(left = lambda df: df.Year0-(window-1))
        .assign(right = lambda df: df.Year0+window)
        .assign(left = lambda df: df.apply(lambda row: max(row.left, row.Yfp), axis=1))
        .assign(right = lambda df: df.apply(lambda row: min(row.right, row.end_year), axis=1))

    ) # latest, editor start in 2014, and have 5 years of service, i.e. 14 15, 16, 17, 18

    print('Editors:', editors.shape) # 5: (15956, 4)
    
    return editors

def getEditorPaper(editors, papers, window):
    # editors: NewAuthorId, issn, left, right
    
    edi_papers = (
        editors.merge(papers, on='NewAuthorId')
        
        .query('Year >= left')
        .query('Year <= right')
        .pipe(printStats, 'Only papers in range')

        .merge(elsevier_journals, on=['JournalId','issn'])
        .pipe(printStats, 'Only elsevier journals')

        .merge(editorials, on='PaperId', how='left').fillna(0)
        .pipe(printStats, 'Shape not change')
        
        .query('Editorial == 0') # filter editorials
        .pipe(printStats, 'Editorials filtered')
    )
    print('Editor papers:', edi_papers.shape)
    
    (
        edi_papers[['NewAuthorId','issn','Year0','PaperId','Year','Editorial']]
        .to_csv(f"../data/EditorJournalPublicationNoEditorial_{window}.csv",
                sep='\t',index=False)
    )
    
    return edi_papers

def getPaperCount(editors, edi_papers):
    
    before = edi_papers[edi_papers.Year <= edi_papers.Year0].groupby(
        ['NewAuthorId','issn','Year0','left','right']).PaperId.nunique().reset_index()

    after = edi_papers[edi_papers.Year > edi_papers.Year0].groupby(
        ['NewAuthorId','issn','Year0','left','right']).PaperId.nunique().reset_index()

    before = before.rename(columns={'PaperId':'BeforeCount'})
    after = after.rename(columns={'PaperId':'AfterCount'})

    print(f'Before: {before.shape}, after: {after.shape}')
    
    paper_count = (
        before.merge(after, on=['NewAuthorId','issn','Year0','left','right'], how='outer')
        .merge(
            editors[['NewAuthorId','issn','Year0','left','right']],
            on=['NewAuthorId','issn','Year0','left','right'], how='outer')
        .fillna(0)

        .assign(BeSpan = lambda df: df.Year0 - df.left + 1)
        .assign(AfSpan = lambda df: df.right - df.Year0)

        .assign(AfAvg = lambda df: df.AfterCount/df.AfSpan)
        .assign(BeAvg = lambda df: df.BeforeCount/df.BeSpan)
    )
    
    print(f'Paper count: {paper_count.shape}')
    
    return paper_count

editors = getEditors("../data/SampleEditors.csv", 5)

edi_papers = getEditorPaper(editors, papers, 5)

paper_count = getPaperCount(editors, edi_papers)

paper_count
