==> figure_3/2.calculate_avg_self_citation_rate.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm

import sys
sys.path.insert(1, '../src')
from utils import priorCount, getCount

## Load data

editors = pd.read_csv("../data/SampleEditors.csv", sep='\t',
                      dtype={'issn':str,'NewAuthorId':int,'start_year':int,'end_year':int})
editors = editors[editors.start_year < 2009]
editors.shape

editor_career = pd.read_csv('../data/EditorCareerDiscipline.csv',sep='\t',
            dtype={'NewAuthorId':int, 'Yfp':int, 'Ylp':int, 'Parent':int})
editor_career.shape

citations = pd.read_csv('../data/EditorCitations.csv',sep='\t',
                       dtype={'NewAuthorId':int,'CitesFrom':int,'BeingCited':int,'Year':int})
citations.shape, citations.NewAuthorId.nunique()

papers = pd.read_csv('../data/EditorPapers.csv',sep='\t',usecols=['NewAuthorId','PaperId'],
                    dtype={'NewAuthorId':int,'PaperId':int,'Year':int, 'JournalId':int})
papers.shape, papers.NewAuthorId.nunique()

%%time
selfCitations = citations.merge(papers, left_on=['CitesFrom','NewAuthorId'], right_on=['PaperId','NewAuthorId'])
print(selfCitations.shape)

selfCitations = selfCitations.drop('PaperId',axis=1)

## Count and prior count

def aggregate(df):
    return df.assign(Count=1).groupby(['NewAuthorId','Year']).Count.sum().reset_index()

citeCount = aggregate(citations)
selfCiteCount = aggregate(selfCitations)
citeCount.shape, selfCiteCount.shape

priorCiteCount = priorCount(citeCount)
priorSelfCiteCount = priorCount(selfCiteCount)

## Editor career

editors = editors.merge(editor_career, on='NewAuthorId')
editors.shape

editors = editors.assign(year0 = editors.start_year-1)
editors = editors.assign(left = editors.year0-10)
editors = editors.assign(right = editors.year0+10)

editors.head()

## Calculate average of 10-year window before and after $\text{year}_0$

%%time
merged = getCount(editors, 'left', priorSelfCiteCount, 'LeftSc')
print(editors.shape, merged.shape)

merged = getCount(merged, 'year0', priorSelfCiteCount, 'Year0Sc')
print(merged.shape)

merged = getCount(merged, 'right', priorSelfCiteCount, 'RightSc')
print(merged.shape)

%%time
merged = getCount(merged, 'left', priorCiteCount, 'LeftIm')
print(merged.shape)

merged = getCount(merged, 'year0', priorCiteCount, 'Year0Im')
print(merged.shape)

merged = getCount(merged, 'right', priorCiteCount, 'RightIm')
print(merged.shape)

merged = merged.assign(BeforeSc = merged.Year0Sc - merged.LeftSc) # self-citation count before year0
merged = merged.assign(AfterSc = merged.RightSc - merged.Year0Sc) # self-citation count after year0

merged = merged.assign(BeforeIm = merged.Year0Im - merged.LeftIm) # impact before year0
merged = merged.assign(AfterIm = merged.RightIm - merged.Year0Im) # impact after year0

merged = merged.assign(BeforeSCPercent = merged.apply(
    lambda x: 0 if x.BeforeIm == 0 else x.BeforeSc/x.BeforeIm, axis=1)) # percentage of self citation before

merged = merged.assign(AfterSCPercent = merged.apply(
    lambda x: 0 if x.AfterIm == 0 else x.AfterSc/x.AfterIm, axis=1)) # percentage of self citation after

merged[['issn','NewAuthorId','BeforeSCPercent','AfterSCPercent']].head()
