==> figure_3/3.match.ipynb <==
import sys
sys.path.insert(1, '../src')
from utils import priorCount, getCount
import pandas as pd
import numpy as np
from tqdm.notebook import tqdm

%%time
editors = pd.read_csv("../data/SampleEditors.csv", sep='\t',
                     dtype={"NewAuthorId":int,"issn":str,"start_year":int,"end_year":int})

ediFirstAff = pd.read_csv('../data/EditorFirstAffRank.csv',sep='\t',dtype={'NewAuthorId':int,'Rank':int})

editor_career = pd.read_csv('../data/EditorCareerDiscipline.csv',sep='\t',
            dtype={'NewAuthorId':int, 'Yfp':int, 'Ylp':int, 'Parent':int})

edi_gender = pd.read_csv('../data/EditorGender.csv',sep='\t',dtype={'NewAuthorId':int,'Gender':str})
print(edi_gender.shape)

# bin affiliation rank into bins
def map_rank(tiers):
    rank_mapping = []
    t = 0
    for i in range(1001):
        if i > tiers[t]:
            t+= 1
        rank_mapping.append({'Rank':i,'Tier':tiers[t]})
    rank_mapping.append({'Rank':1001,'Tier':1001})
    rank_mapping = pd.DataFrame(rank_mapping)

    return rank_mapping

rank_mapping = map_rank([20, 50, 100, 300, 600, 1000])

## Editors

%%time
editors = pd.read_csv("../data/SampleEditors.csv", sep='\t',
                     dtype={"NewAuthorId":int,"issn":str,"start_year":int,"end_year":int})

editors = editors.merge(editor_career, on='NewAuthorId')
editors = editors.merge(ediFirstAff, on='NewAuthorId')
editors = editors.merge(edi_gender, on='NewAuthorId')
editors = editors.merge(rank_mapping, on='Rank')
editors = editors.drop('Rank',axis=1).drop_duplicates()

print(editors.shape, editors.NewAuthorId.nunique())
editors = editors.rename(columns={'Ylp':'Eylp','Yfp':'Eyfp','NewAuthorId':'EditorsNewId'})
editors = editors.assign(EditorYear0=editors.start_year-1)

editors.head()

extended = []
for delta in range(-3, 4): # year-of-first-publication can be differed by at most 3 years
    extended.append(editors.assign(Ayfp=editors.Eyfp+delta))

editors = pd.concat(extended, ignore_index=True, sort=False)
editors.shape, editors.EditorsNewId.nunique()

editors = editors.assign(age=editors.EditorYear0 - editors.Eyfp)
editors = editors.assign(age_range=editors.age.apply(lambda x: max(1, round(x*0.1) )) ) # differ by minimum 1 year

%%time
extended = pd.DataFrame()
for age in tqdm(range(int(editors.age_range.max()) +1)):
    temp = editors[editors.age_range == age]

    for diff in range(-age, age+1):
        newtemp = temp.assign(AuthorYear0=temp.Ayfp + diff + temp.age)

        # filter out two cases where difference in end_year is 4
        newtemp = newtemp.assign(end_diff=(newtemp.AuthorYear0 - newtemp.EditorYear0).apply(lambda x: abs(x)))
        newtemp = newtemp[newtemp.end_diff <= 3]

        extended = extended.append(newtemp, ignore_index=True, sort=False)
extended = extended.drop(['end_diff', 'age', 'age_range'], axis=1)

editors = extended.copy()
print(editors.shape, editors.EditorsNewId.nunique())

editors.head()

## Authors

author_career = pd.read_csv('../data/figure_2/AuthorCareer.csv',sep='\t',
            dtype={'NewAuthorId':int, 'Yfp':int, 'Ylp':int, 'Parent':int})
print(author_career.shape)

# load rank of first aff
firstAff = pd.read_csv('../data/figure_2/AuthorFirstAffRank.csv',sep='\t',
                      usecols=['NewAuthorId', 'Rank'], dtype={'NewAuthorId':int,'Rank':int})
print(firstAff.shape)

author_gender = pd.read_csv('../data/figure_2/AuthorGender.csv',sep='\t',
                              dtype={'NewAuthorId':int,'gender':str})
print(author_gender.shape)

%%time
authors = author_career.merge(firstAff, on='NewAuthorId').merge(author_gender, on='NewAuthorId')
print(authors.shape)

%%time
authors = authors.merge(rank_mapping, on='Rank').drop('Rank',axis=1).drop_duplicates()
print(authors.shape)

authors = authors.rename(columns={'Yfp':'Ayfp','Ylp':'Aylp'})

## Potential matches

At this stage, match editors and authors on their discipline, gender, rank of first affiliation, and year of first publication.

%%time
matched = editors.merge(authors, on=['Ayfp','Parent','gender','Tier'])
matched = matched[~matched.NewAuthorId.isin(editors.EditorsNewId)]

print(matched.shape, end='\t')

## make sure that start and end publishing before and after year0 ##
matched = matched.assign(lower = np.minimum(matched.EditorYear0, matched.AuthorYear0))
matched = matched.assign(upper = np.maximum(matched.EditorYear0, matched.AuthorYear0))

matched = matched[matched.Aylp >= matched.upper]
matched = matched[matched.Ayfp <= matched.lower]
matched = matched[matched.Eylp >= matched.upper]
matched = matched[matched.Eyfp <= matched.lower]

matched = matched.drop(['lower', 'upper'], axis=1)
## make sure that start and end publishing before and after year0 ##

matched = matched.drop_duplicates(subset=['EditorsNewId','issn','NewAuthorId','AuthorYear0'])
matched.shape # 315698

editors.head()

matched = (
    matched.merge(
        editors[['EditorsNewId','issn']]
        .assign(IsEditor = 1)
        .rename(columns={'EditorsNewId':'NewAuthorId'}),
        on=['NewAuthorId','issn'], how='left')
    .fillna(0)
    .query('IsEditor == 0')
    .drop('IsEditor', axis=1)
) # make sure that an author matched to (e, j) is not an editor of j
matched.shape

matched.head() # potential pairs of matched editors and authors

## Paper count and paper citation

jCiteCount = pd.read_csv("../data/figure_2/JournalCitationCount.csv",sep='\t',
                        dtype={'NewAuthorId':int, 'issn':str, 'Year':int, 'Count':int})
jCiteCount.shape

jPaperCount = pd.read_csv("../data/figure_2/JournalPaperCount.csv",sep='\t',
                        dtype={'NewAuthorId':int, 'issn':str, 'Year':int, 'Count':int})
jPaperCount.shape

priorCite = priorCount(jCiteCount)

priorPub = priorCount(jPaperCount)

def mergeCount(df, count_df, left_on, col):
    df = df.merge(count_df, how='left',
                    left_on=left_on,
                    right_on=['NewAuthorId','Year','issn']).fillna(0)
    df = df.rename(columns={'Count':col})
    df = df.drop(['NewAuthorId','Year'],axis=1)
    
    return df

def mergePrior(df, prior_df, year_col, id_col, val_col):
    df = df.rename(columns={id_col: 'NewAuthorId'})
    
    df = getCount(df, year_col, prior_df, val_col)
    
    df = df.rename(columns={'NewAuthorId':id_col})
    
    return df

matchedCount = matched.drop(['Parent','gender','Tier','Ayfp'],axis=1)

matchedCount = matchedCount.rename(columns={'NewAuthorId':'AuthorsNewId'})

# author citation count in the editor's journal in year0
matchedCount = mergeCount(matchedCount, jCiteCount, ['AuthorsNewId','AuthorYear0','issn'], 'AutCiteCount')

# editor citation count in the editor's journal in year0
matchedCount = mergeCount(matchedCount, jCiteCount, ['EditorsNewId','EditorYear0','issn'], 'EdiCiteCount')

# author paper count in the editor's journal in year0
matchedCount = mergeCount(matchedCount, jPaperCount, ['AuthorsNewId','AuthorYear0','issn'], 'AutPubCount')

# editor paper count in the editor's journal in year0
matchedCount = mergeCount(matchedCount, jPaperCount, ['EditorsNewId','EditorYear0','issn'], 'EdiPubCount')

# author total citation count in the editor's journal up until year0 
matchedCount = mergePrior(matchedCount, priorCite, 'AuthorYear0', 'AuthorsNewId', 'AutCitePrior')

# author total paper count in the editor's journal up until year0
matchedCount = mergePrior(matchedCount, priorCite, 'EditorYear0', 'EditorsNewId', 'EdiCitePrior')

# editor total citation count in the editor's journal up until year0
matchedCount = mergePrior(matchedCount, priorPub, 'AuthorYear0', 'AuthorsNewId', 'AutPubPrior')

# editor total paper count in the editor's journal up until year0
matchedCount = mergePrior(matchedCount, priorPub, 'EditorYear0', 'EditorsNewId', 'EdiPubPrior')

matchedCount.head()

def addOne(df, outcome, typ):
    # avoid divide by 0, does not affect outcome
    edi = 'Edi'+outcome+typ
    aut = 'Aut'+outcome+typ
    df = df.assign(Ec=df.apply(lambda row: row[edi]+1 if row[edi] == 0 else row[edi], axis=1))
    df = df.assign(Ac=df.apply(lambda row: row[aut]+1 if row[edi] == 0 else row[aut], axis=1))
    
    return df
    
def _filterMatched(matched, threshold):

    matched['diff'] = matched.apply(lambda row: abs(row['Ec']-row['Ac'])/row['Ec'], axis=1)
    filtered = matched[(matched['diff'] <= threshold)]
    matched = filtered.drop(['diff','Ac','Ec'], axis=1)
    
    return matched
    
def filterMatched(matched, outcome, threshold):
    
    matched = addOne(matched, outcome, 'Prior')
    print(matched.shape)
    
    matched = _filterMatched(matched, threshold)
    
    print(matched.shape)
    
    matched = addOne(matched, outcome, 'Count')
    print(matched.shape)
    
    matched = _filterMatched(matched, threshold)
    print(matched.shape)

    return matched

%%time
citeMatched = filterMatched(matchedCount, 'Cite', 0.1)

%%time
pubMatched = filterMatched(matchedCount, 'Pub', 0.1)
